+++ CHANGELOG.rst
@@ -40,6 +40,9 @@
 
 * removed the message type 'sysex_end'. It had no use and only caused problems.
 
+* Message() no longer accepts a status byte as its first argument. (This was
+  only meant to be used internally.)
+
 
 1.0.2 - 2013-07-31
 -------------------